% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCat.r
\name{getCat}
\alias{getCat}
\title{Get cat function}
\usage{
getCat(cat = TRUE)
}
\arguments{
\item{cat}{Logical. If true a cat is generated; if false a dog is generated.}
}
\description{
This function will get a cat for you.
}
\examples{
getCat()
}
\keyword{cats}
